requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.3
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.1
twilio==8.0.0
requests==2.31.0
gunicorn==21.2.0
captcha==0.7.1